+++ CHANGELOG.md
@@ -62,6 +62,7 @@ All notable changes to this project will be documented in this file.
 - Bug where liking a repost also liked the original post
 - Incorrect use of `original_post_id` in repost like logic
 - Import bug for all `models/`, `utils/`, `routers/` and `main.py` files
+- Complete fixed bug that prohibited backend to start up. 
 
 ## [1.0.10]